requirements.txt
PyQt5==5.9.2
pyserial
pandas
fbs